refactor: address orange-priority architectural debt
- config: validate SECRET_KEY and DATABASE_URL in production environment to
  prevent accidental deploy with insecure defaults
- models: add XOR CHECK constraints (num_nonnulls = 1) on vehicle_id/machine_id
  in maintenance_records, work_orders, and fleet_assignments; enforced at DB level
- models: fix float -> Decimal type annotation for cost/amount fields backed by
  Numeric(12,2) columns (prevents float precision loss on monetary values)
- migration n9i0j1k2l3m4: applies the three XOR constraints to existing tables
- useList: increase default page size from 100 to 500 to prevent silently
  truncated dropdowns for larger fleets

+++ backend/app/models/quote.py
@@ -2,6 +2,7 @@
 
 import uuid
 from datetime import date
+from decimal import Decimal
 from enum import Enum
 
 from sqlalchemy import Date, ForeignKey, Numeric, String, Text, UUID
@@ -33,7 +34,7 @@ class Quote(Base, TimestampMixin):
         UUID(as_uuid=True), ForeignKey("work_orders.id", ondelete="SET NULL"), nullable=True, index=True
     )
     description: Mapped[str] = mapped_column(Text, nullable=False)
-    amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
+    amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     status: Mapped[str] = mapped_column(String(50), default=EstadoPresupuesto.PENDIENTE, nullable=False, index=True)
     valid_until: Mapped[date | None] = mapped_column(Date)
     notes: Mapped[str | None] = mapped_column(Text)
